+++ lib/default/TasmotaSerial-3.6.0/src/TasmotaSerial.cpp
@@ -46,9 +46,11 @@ static uint32_t tasmota_serial_uart_bitmap = 0;      // Assigned UARTs
 
 TasmotaSerial::TasmotaSerial(int receive_pin, int transmit_pin, int hardware_fallback, int nwmode, int buffer_size) {
   m_valid = false;
+  m_tx_enable_valid = false;
   m_hardserial = false;
   m_hardswap = false;
   m_overflow = false;
+  m_data_bits = 8;
   m_stop_bits = 1;
   m_nwmode = nwmode;
   serial_buffer_size = buffer_size;
@@ -122,7 +124,23 @@ TasmotaSerial::~TasmotaSerial(void) {
 }
 
 bool TasmotaSerial::isValidGPIOpin(int pin) {
+#ifdef ESP8266
   return (pin >= -1 && pin <= 5) || (pin >= 12 && pin <= 15);
+#endif
+#ifdef ESP32
+  return GPIO_IS_VALID_OUTPUT_GPIO(pin);
+#endif
+}
+
+void TasmotaSerial::setTransmitEnablePin(int tx_enable_pin) {
+  if ((tx_enable_pin > -1) && isValidGPIOpin(tx_enable_pin)) {
+    m_tx_enable_valid = true;
+    m_tx_enable_pin = tx_enable_pin;
+    pinMode(m_tx_enable_pin, OUTPUT);
+    digitalWrite(m_tx_enable_pin, LOW);
+  } else {
+    m_tx_enable_valid = false;
+  }
 }
 
 #ifdef ESP32
@@ -239,12 +257,17 @@ bool TasmotaSerial::begin(uint32_t speed, uint32_t config) {
 //    Serial.printf("TSR: Using UART%d\n", m_uart);
 #endif  // ESP32
   } else {
+    // #define UART_NB_BIT_5         0B00000000
+    // #define UART_NB_BIT_6         0B00000100
+    // #define UART_NB_BIT_7         0B00001000
+    // #define UART_NB_BIT_8         0B00001100
+    m_data_bits = 5 + ((config &0x0C) >> 2);
     // Software serial fakes two stop bits if either stop bits is 2 or parity is not None
     // #define UART_NB_STOP_BIT_0    0B00000000
     // #define UART_NB_STOP_BIT_1    0B00010000
     // #define UART_NB_STOP_BIT_15   0B00100000
     // #define UART_NB_STOP_BIT_2    0B00110000
-    m_stop_bits = ((config &0x30) >> 5) +1;
+    m_stop_bits = 1 + ((config &0x30) >> 5);
     // #define UART_PARITY_NONE      0B00000000
     // #define UART_PARITY_EVEN      0B00000010
     // #define UART_PARITY_ODD       0B00000011
@@ -368,9 +391,9 @@ int TasmotaSerial::available(void) {
   }
 }
 
-#define TM_SERIAL_WAIT_SND { while (ESP.getCycleCount() < (wait + start)) if (!m_high_speed) optimistic_yield(1); wait += m_bit_time; } // Watchdog timeouts
+#define TM_SERIAL_WAIT_SND      { while (ESP.getCycleCount() < (wait + start)) if (!m_high_speed) optimistic_yield(1); wait += m_bit_time; } // Watchdog timeouts
 #define TM_SERIAL_WAIT_SND_FAST { while (ESP.getCycleCount() < (wait + start)); wait += m_bit_time; }
-#define TM_SERIAL_WAIT_RCV { while (ESP.getCycleCount() < (wait + start)); wait += m_bit_time; }
+#define TM_SERIAL_WAIT_RCV      { while (ESP.getCycleCount() < (wait + start)); wait += m_bit_time; }
 #define TM_SERIAL_WAIT_RCV_LOOP { while (ESP.getCycleCount() < (wait + start)); }
 
 void IRAM_ATTR TasmotaSerial::_fast_write(uint8_t b) {
@@ -379,7 +402,7 @@ void IRAM_ATTR TasmotaSerial::_fast_write(uint8_t b) {
   // Start bit;
   digitalWrite(m_tx_pin, LOW);
   TM_SERIAL_WAIT_SND_FAST;
-  for (uint32_t i = 0; i < 8; i++) {
+  for (uint32_t i = 0; i < m_data_bits; i++) {
     digitalWrite(m_tx_pin, (b & 1) ? HIGH : LOW);
     TM_SERIAL_WAIT_SND_FAST;
     b >>= 1;
@@ -392,15 +415,20 @@ void IRAM_ATTR TasmotaSerial::_fast_write(uint8_t b) {
 }
 
 size_t TasmotaSerial::write(uint8_t b) {
+  if (!m_hardserial && (-1 == m_tx_pin)) { return 0; }
+
+  if (m_tx_enable_valid) {
+    digitalWrite(m_tx_enable_pin, HIGH);
+  }
+  size_t size = 0;
   if (m_hardserial) {
 #ifdef ESP8266
-    return Serial.write(b);
+    size = Serial.write(b);
 #endif  // ESP8266
 #ifdef ESP32
-    return TSerial->write(b);
+    size = TSerial->write(b);
 #endif  // ESP32
   } else {
-    if (-1 == m_tx_pin) return 0;
     if (m_high_speed) {
       cli();  // Disable interrupts in order to get a clean transmit
       _fast_write(b);
@@ -412,7 +440,7 @@ size_t TasmotaSerial::write(uint8_t b) {
       // Start bit;
       digitalWrite(m_tx_pin, LOW);
       TM_SERIAL_WAIT_SND;
-      for (uint32_t i = 0; i < 8; i++) {
+      for (uint32_t i = 0; i < m_data_bits; i++) {
         digitalWrite(m_tx_pin, (b & 1) ? HIGH : LOW);
         TM_SERIAL_WAIT_SND;
         b >>= 1;
@@ -424,25 +452,29 @@ size_t TasmotaSerial::write(uint8_t b) {
         TM_SERIAL_WAIT_SND;
       }
     }
-
-    return 1;
+    size = 1;
   }
+  if (m_tx_enable_valid) {
+    digitalWrite(m_tx_enable_pin, LOW);
+  }
+  return size;
 }
 
 void IRAM_ATTR TasmotaSerial::rxRead(void) {
   uint32_t m_out_pos_fixed = m_out_pos;
   if (!m_nwmode) {
     int32_t loop_read = m_very_high_speed ? serial_buffer_size : 1;
+    uint32_t bit_mask = 0x01 << (m_data_bits -1);
     // Advance the starting point for the samples but compensate for the
     // initial delay which occurs before the interrupt is delivered
     uint32_t wait = m_bit_start_time;
     uint32_t start = ESP.getCycleCount();
-    while (loop_read-- > 0) {    // try to receveive all consecutive bytes in a row
+    while (loop_read-- > 0) {    // try to receive all consecutive bytes in a row
       uint32_t rec = 0;
-      for (uint32_t i = 0; i < 8; i++) {
+      for (uint32_t i = 0; i < m_data_bits; i++) {
         TM_SERIAL_WAIT_RCV;
         rec >>= 1;
-        if (digitalRead(m_rx_pin)) rec |= 0x80;
+        if (digitalRead(m_rx_pin)) rec |= bit_mask;
       }
       // Store the received value in the buffer unless we have an overflow
       uint32_t next = (m_in_pos + 1) % serial_buffer_size;
@@ -485,6 +517,7 @@ void IRAM_ATTR TasmotaSerial::rxRead(void) {
     // it gets set even when interrupts are disabled
     GPIO_REG_WRITE(GPIO_STATUS_W1TC_ADDRESS, 1 << m_rx_pin);
   } else {
+    // Currently supports 8-bit data only
     uint32_t diff;
     uint32_t level;