+++ lib/default/TasmotaSerial-3.6.0/src/TasmotaSerial.h
@@ -39,6 +39,7 @@ class TasmotaSerial : public Stream {
   public:
     TasmotaSerial(int receive_pin, int transmit_pin, int hardware_fallback = 0, int nwmode = 0, int buffer_size = TM_SERIAL_BUFFER_SIZE);
     virtual ~TasmotaSerial();
+    void setTransmitEnablePin(int tx_enable_pin);
 
     size_t setRxBufferSize(size_t size);
     size_t getRxBufferSize() { return serial_buffer_size; }
@@ -80,6 +81,8 @@ class TasmotaSerial : public Stream {
     // Member variables
     int m_rx_pin;
     int m_tx_pin;
+    int m_tx_enable_pin;
+    uint32_t m_data_bits;
     uint32_t m_stop_bits;
     uint32_t ss_byte;
     uint32_t ss_bstart;
@@ -91,6 +94,7 @@ class TasmotaSerial : public Stream {
     uint32_t m_out_pos;
     uint32_t serial_buffer_size = TM_SERIAL_BUFFER_SIZE;
     bool m_valid;
+    bool m_tx_enable_valid;
     bool m_nwmode;
     bool m_hardserial;
     bool m_hardswap;