+++ lib/lib_basic/TasmotaModbus-3.6.0/src/TasmotaModbus.cpp
@@ -27,8 +27,9 @@ enum LoggingLevels {LOG_LEVEL_NONE, LOG_LEVEL_ERROR, LOG_LEVEL_INFO, LOG_LEVEL_D
 
 //#define TASMOTAMODBUSDEBUG
 
-TasmotaModbus::TasmotaModbus(int receive_pin, int transmit_pin) : TasmotaSerial(receive_pin, transmit_pin, 1)
+TasmotaModbus::TasmotaModbus(int receive_pin, int transmit_pin, int tx_enable_pin) : TasmotaSerial(receive_pin, transmit_pin, 1)
 {
+  setTransmitEnablePin(tx_enable_pin);
   mb_address = 0;
 }
 
@@ -91,7 +92,7 @@ uint8_t TasmotaModbus::Send(uint8_t device_address, uint8_t function_code, uint1
   }
   else if ((function_code == 5) || (function_code == 6))
   {
-    if (write_data == NULL) 
+    if (write_data == NULL)
     {
       free(frame);
       return 13; // Register data not specified
@@ -108,10 +109,10 @@ uint8_t TasmotaModbus::Send(uint8_t device_address, uint8_t function_code, uint1
   {
     frame[framepointer++] = (uint8_t)(count >> 8);   // MSB
     frame[framepointer++] = (uint8_t)(count);        // LSB
-    
+
     frame[framepointer++] = byte_count;
 
-    if (write_data == NULL) 
+    if (write_data == NULL)
     {
       free(frame);
       return 13; // Register data not specified
@@ -126,7 +127,7 @@ uint8_t TasmotaModbus::Send(uint8_t device_address, uint8_t function_code, uint1
       frame[framepointer++] = (uint8_t)(write_data[bytepointer/2] >> (bytepointer % 2 ? 0 : 8));  // MSB, LSB, MSB ....
     }
   }
-  else 
+  else
   {
     free(frame);
     return 1; // Wrong function code
@@ -136,7 +137,7 @@ uint8_t TasmotaModbus::Send(uint8_t device_address, uint8_t function_code, uint1
   frame[framepointer++] = (uint8_t)(crc);
   frame[framepointer++] = (uint8_t)(crc >> 8);
 
-#ifdef TASMOTAMODBUSDEBUG  
+#ifdef TASMOTAMODBUSDEBUG
   uint8_t *buf;
   uint16_t bufsize=(framepointer + 1) * 3;
   buf = (uint8_t *)malloc(bufsize);
@@ -147,7 +148,7 @@ uint8_t TasmotaModbus::Send(uint8_t device_address, uint8_t function_code, uint1
   AddLog(LOG_LEVEL_DEBUG_MORE, PSTR("MBS: Serial Send: %s"), buf);
   free(buf);
 #endif
-  
+
   flush();
   write(frame, framepointer);
   free(frame);
@@ -213,7 +214,7 @@ uint8_t TasmotaModbus::ReceiveBuffer(uint8_t *buffer, uint8_t register_count, ui
                       // 10 = Gateway Path Unavailable
                       // 11 = Gateway Target device failed to respond
   }
-  
+
   if (mb_len < 6) { return 7; }  // 7 = Not enough data
 
 /*